requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
transformers>=4.36.2
torch>=2.1.2
sentencepiece>=0.1.99
protobuf>=3.20.0
matplotlib>=3.7.0
seaborn>=0.12.0
fpdf2>=2.7.6
supabase>=2.3.4
postgrest>=0.16.0
gotrue>=2.4.2
requests>=2.31.0
python-dotenv>=1.0.0 